requirements.txt
transformers==4.11
datasets
seqeval
wandb
plotly
accelerate
more_itertools
tensorboard
torchinfo